Add `.pylintrc` to ensure `rip_scripts` is actually scanned in VSCode.
It is absolutely, completely frustrating difficult that there is no actual way in `.pylintrc` to just say "this is my Python source directory, scan that". Hell, they don't even document `.pylintrc` - you're expected to use Pylint itself to generate a sample file with a bunch of inane settings and then customize what you like out of it. This is not Google-friendly.

For those Googling "how to make pylint scan my imports correctly" the answer is this commit specifically. You have to set an `init-hook` that adds the path of your Python source directory to `sys.path`, and then it will magically work.